Refactor to use DisplaySelectors component
