+++ client/src/components/ProTx/DashboardDisplay.module.scss
@@ -1,0 +1,12 @@
+.root {
+  flex-grow: 1;
+  display: flex;
+  flex-direction: column;
+  font: 16px Roboto, sans-serif;
+}
+
+.error {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
