--- client/src/components/ProTx/DisplaySelectors.css
@@ -1,4 +1,0 @@
-.display-selectors-root {
-    height: 38px;
-    background-color: #f60;
-}
